create map-types file
