+++ src/pages/components/map-related/MapWithLocationsAndPath.tsx
@@ -1,29 +1,15 @@
 import { MapContainer } from 'react-leaflet'
 import { Point } from 'src/pages/timeBasedMap'
 import { MapContent } from './MapContent'
-
-import { VehicleLocation } from 'src/model/vehicleLocation'
+import { MapProps } from './map-types'
 import { useCallback, useState } from 'react'
 import { Button } from 'antd'
 import { ExpandAltOutlined } from '@ant-design/icons'
-import { BusStop } from 'src/model/busStop'
 import '../../Map.scss'
 
 export const position: Point = {
   loc: [32.3057988, 34.85478613], // arbitrary default value... Netanya - best city to live & die in
   color: 0,
-}
-
-export interface Path {
-  locations: VehicleLocation[]
-  lineRef: number
-  operator: number
-  vehicleRef: number
-}
-
-export interface MapProps {
-  positions: Point[]
-  plannedRouteStops: BusStop[]
 }
 
 export function MapWithLocationsAndPath({ positions, plannedRouteStops }: MapProps) {
@@ -45,22 +31,3 @@ export function MapWithLocationsAndPath({ positions, plannedRouteStops }: MapPro
     </div>
   )
 }
-
-// export function RecenterOnDataChange({ positions, plannedRouteStops }: MapProps) {
-//   const map = useMap()
-
-//   useEffect(() => {
-//     const positionsSum = positions.reduce(
-//       (acc, { loc }) => [acc[0] + loc[0], acc[1] + loc[1]],
-//       [0, 0],
-//     )
-//     const mean: LatLngTuple = [
-//       positionsSum[0] / positions.length || position.loc[0],
-//       positionsSum[1] / positions.length || position.loc[1],
-//     ]
-
-//     map.setView(mean, map.getZoom(), { animate: true })
-//   }, [positions, plannedRouteStops])
-
-//   return null
-// }
