+++ src/pages/components/map-related/useRecenterOnDataChange.ts
@@ -1,7 +1,8 @@
 import { useEffect } from 'react'
 import { LatLngTuple } from 'leaflet'
 import { useMap } from 'react-leaflet'
-import { MapProps, position } from './MapWithLocationsAndPath'
+import { position } from './MapWithLocationsAndPath'
+import { MapProps } from './map-types'
 
 export function useRecenterOnDataChange({ positions, plannedRouteStops }: MapProps) {
   const map = useMap()
@@ -13,16 +14,8 @@ export function useRecenterOnDataChange({ positions, plannedRouteStops }: MapPro
     positionsSum[0] / positions.length || position.loc[0],
     positionsSum[1] / positions.length || position.loc[1],
   ]
+  console.log('mean: ', mean)
   useEffect(() => {
-    // const positionsSum = positions.reduce(
-    //   (acc, { loc }) => [acc[0] + loc[0], acc[1] + loc[1]],
-    //   [0, 0],
-    // )
-    // const mean: LatLngTuple = [
-    //   positionsSum[0] / positions.length || position.loc[0],
-    //   positionsSum[1] / positions.length || position.loc[1],
-    // ]
-
     map.setView(mean, map.getZoom(), { animate: true })
   }, [positions, plannedRouteStops])
 }
